+++ Project/CODE/driver/BaseDriver.hpp
@@ -8,11 +8,11 @@
 class BaseDriver {
  public:
     struct WheelSpeed {
-        float L1, L2, R1, R2;
+        double L1, L2, R1, R2;
         void setZero() { L1 = L2 = R1 = R2 = 0; }
     };
     struct BaseSpeed {
-        float x, y, yaw;
+        double x, y, yaw;
     };
 
     struct CmdVelAcc {
@@ -54,9 +54,9 @@ class BaseDriver {
 
  public:
     // 车底盘中心到轮子中心的距离的分量, 单位是m
-    static constexpr float r_x = 0.198 / 2, r_y = (0.21 - 0.0325) / 2;
+    static constexpr double r_x = 0.198 / 2, r_y = (0.21 - 0.0325) / 2;
 
-    static inline WheelSpeed calc_vel(float x, float y, float yaw) {
+    static inline WheelSpeed calc_vel(double x, double y, double yaw) {
         WheelSpeed res;
         res.L1 = x - y - yaw * (r_x + r_y);
         res.L2 = x + y - yaw * (r_x + r_y);
@@ -65,7 +65,7 @@ class BaseDriver {
         return res;
     }
 
-    static inline BaseSpeed calc_vel(float L1, float L2, float R1, float R2) {
+    static inline BaseSpeed calc_vel(double L1, double L2, double R1, double R2) {
         BaseSpeed res;
         res.x = (L1 + L2 + R1 + R2) / 4;
         res.y = (-L1 + L2 + R1 - R2) / 4;
@@ -73,16 +73,16 @@ class BaseDriver {
         return res;
     }
 
-    inline void cmd_vel(float x, float y, float yaw) { _wheelSpeedLoader.store(calc_vel(x, y, yaw)); }
+    inline void cmd_vel(double x, double y, double yaw) { _wheelSpeedLoader.store(calc_vel(x, y, yaw)); }
     inline void cmd_vel(const BaseSpeed& base) { cmd_vel(base.x, base.y, base.yaw); }
-    inline void cmd_vel(float L1, float L2, float R1, float R2) { _wheelSpeedLoader.store({L1, L2, R1, R2}); }
+    inline void cmd_vel(double L1, double L2, double R1, double R2) { _wheelSpeedLoader.store({L1, L2, R1, R2}); }
     inline void cmd_vel(const WheelSpeed& wheel) { _wheelSpeedLoader.store(wheel); }
 
-    inline void cmd_vel_acc(float vX, float vY, float vYaw, float aX, float aY, float aYaw, uint64_t timestamp_us) {
+    inline void cmd_vel_acc(double vX, double vY, double vYaw, double aX, double aY, double aYaw, uint64_t timestamp_us) {
         _cmdVelAccLoader.store({timestamp_us, true, {vX, vY, vYaw}, {aX, aY, aYaw}});
     }
 
-    inline void get_vel(float L1, float L2, float R1, float R2) { _baseSpeed = calc_vel(L1, L2, R1, R2); }
+    inline void get_vel(double L1, double L2, double R1, double R2) { _baseSpeed = calc_vel(L1, L2, R1, R2); }
     inline void get_vel(WheelSpeed& wheel) { get_vel(wheel.L1, wheel.L2, wheel.R1, wheel.R2); }
 
     inline void setControlState(bool L1, bool L2, bool R1, bool R2) { _controlStateLoader.emplace(L1, L2, R1, R2); }
